OBJECTS-768: fix tests
